fix os name, make dependency script work on ubuntu

+++ dependencies.sh
@@ -1,17 +1,27 @@
 #!/bin/bash
 set -e
 
-# @TODO Only on macos
-
 function run() { 
   set -x
   "$@" 
   { set +x;   } 2> /dev/null
 }
 
+script_dir=$( cd -- "$( dirname -- "${BASH_SOURCE[0]}" )" &> /dev/null && pwd )
 
-run brew update
-run brew install ccache openssl@3 zlib zstd ncurses expat xerces-c freetype xz lz4 libx11 libxml2 libxpm libxft sqlite
-run brew reinstall cmake
+source ${script_dir}/detect_os.sh
+echo "OS: ${os_name}"
 
-brew unlink r || true
+if [ $os == "ubuntu" ]; then
+  echo "Installing dependencies for Ubuntu"
+elif [ $os == "macos" ]; then
+  echo "Installing dependencies for macOS"
+  run brew update
+  run brew install ccache openssl@3 zlib zstd ncurses expat xerces-c freetype xz lz4 libx11 libxml2 libxpm libxft
+  run brew reinstall cmake
+
+  brew unlink r || true
+else
+  echo "Only Ubuntu and macOS are supported. Exiting."
+  exit 1
+fi
